requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
emoji==2.14.1
fastapi==0.115.11
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.2.0
networkx==3.4.2
numpy==2.2.4
protobuf==6.30.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
requests==2.32.3
setuptools==78.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stanza==1.10.1
starlette==0.46.0
sympy==1.13.1
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.0
pytest==8.3.5